requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
opencv-python-headless==4.8.1.78
PyMuPDF==1.23.8
pyzbar==0.1.9
requests==2.31.0
Pillow==10.0.1
numpy==1.24.3
gunicorn==20.1.0
pytesseract==0.3.10
easyocr==1.7.0